+++ plugins/rules.go
@@ -64,7 +64,7 @@ func (a *Agreement) addAgreementRole(roleID string, msg *dg.MessageCreate) (stri
 
 	// Make sure the bot owner is executing the command
 	if msg.Author.ID != ownerID {
-		return "This command is for the bot owner only ::", nil
+		return "This command is for the bot owner only :rage:", nil
 	}
 
 	// Make sure the string passed is the role ID